@@ -6,26 +6,26 @@ publish_to: "none"
 version: 1.1.3+3
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  bubble_bottom_bar: ^1.2.0
-  cupertino_icons: ^1.0.2
-  date_format: ^2.0.2
-  flutter_local_notifications: ^8.2.0
-  flutter_native_timezone: ^1.0.10
-  flutter_slidable: ^0.5.7
-  flutter_svg: ^0.19.3
-  font_awesome_flutter: ^9.1.0
-  get: ^4.3.8
+  bubble_bottom_bar: ^2.0.0
+  cupertino_icons: ^1.0.4
+  date_format: ^2.0.6
+  flutter_local_notifications: ^9.5.3+1
+  flutter_native_timezone: ^2.0.0
+  flutter_slidable: ^1.3.0
+  flutter_svg: ^1.1.0
+  font_awesome_flutter: ^10.1.0
+  get: ^4.6.5
   get_storage: ^2.0.3
-  google_fonts: ^1.1.1
-  hive: ^2.0.4
+  google_fonts: ^3.0.1
+  hive: ^2.2.1
   intl: ^0.17.0
-  lottie: ^1.1.0
-  path_provider: ^2.0.1
+  lottie: ^1.3.0
+  path_provider: ^2.0.10
   timezone: ^0.8.0
 
 dev_dependencies:
