@@ -15,38 +15,38 @@ import 'package:date_format/date_format.dart';
 import 'package:taskly/app/theme/text_theme.dart';
 
 class HomeController extends GetxController {
-  // bottom nav current index.
+  /// [ bottom nav current index ]
   var _currentIndex = 0.obs;
   get currentIndex => this._currentIndex.value;
   set currentIndex(index) => this._currentIndex.value = index;
 
-  // Hive DB var
+  /// [ Hive DB var ]
   String taskBox = 'tasks';
   int taskIds = 0;
 
-  // instance of Get Storage
+  /// [ instance of Get Storage ]
   final userData = GetStorage();
 
-  // variable for expansionTile
+  /// [ variable for expansionTile ]
   bool isExpanded = false;
 
-  // bools for dashboard view
+  /// [ bools for dashboard view ]
   bool isCurrentTaskPresent = false;
   bool isUpcommingTaskPresent = false;
-  Task currentTask;
-  Task upcomingTask;
+  Task? currentTask;
+  Task? upcomingTask;
 
-  // controllers and var for bottomSheet TextFeilds
-  TextEditingController titleController;
-  TextEditingController descController;
-  TextEditingController dateController;
-  TextEditingController timeController;
-  String selectedIcon;
-  DateTime selectedDate;
-  TimeOfDay selectedTime;
-  String setTime, setDate;
-  String hour, minute, time;
-  bool isRepeat = false;
+  /// [ controllers and var for bottomSheet TextFeilds ]
+  TextEditingController? titleController;
+  TextEditingController? descController;
+  TextEditingController? dateController;
+  TextEditingController? timeController;
+  String? selectedIcon;
+  DateTime? selectedDate;
+  late TimeOfDay selectedTime;
+  String? setTime, setDate;
+  String? hour, minute, time;
+  bool? isRepeat = false;
   List<String> icons = [
     'assets/icons/alarm-clock.svg',
     'assets/icons/breakfast.svg',
@@ -60,63 +60,63 @@ class HomeController extends GetxController {
     'assets/icons/travel.svg',
   ];
 
-  // allTasks slidable
-  SlidableController slideC;
+  /// [ allTasks slidable ]
+  SlidableController? slideC;
 
-  // the list of screens switched by bottom navBar
+  /// [ the list of screens switched by bottom navBar ]
   final List<Widget> homeViews = [
     DashboardView(),
     TodayTaskView(),
   ];
 
-  // temp var to create each task
-  Task tempTask;
+  /// [ temp var to create each task ]
+  Task? tempTask;
 
-  // task lists
-  List<Task> allTasks = [];
-  List<Task> commingTasks = [];
-  List<Task> pastTasks = [];
-  List<Task> todayTasks = [];
+  /// [ task lists ]
+  List<Task?> allTasks = [];
+  List<Task?> commingTasks = [];
+  List<Task?> pastTasks = [];
+  List<Task?> todayTasks = [];
 
-  // userData
-  String userName;
-  bool isMale = false;
+  /// [ userData ]
+  String? userName;
+  bool? isMale = false;
 
-  // instance of notification Plugin
-  NotificationPlugin nPlugin;
+  /// [ instance of notification Plugin ]
+  late NotificationPlugin nPlugin;
 
-  // function to return correct view on bottom navBar switch
+  /// [ function to return correct view on bottom navBar switch ]
   Widget navBarSwitcher() {
     return homeViews.elementAt(currentIndex);
   }
 
-  // ExpandedContainer
+  /// [ ExpandedContainer ]
   void onExpand(bool value) {
     isExpanded = value;
     update();
   }
 
-  // change icon in bottomSheet
-  changeIcon(String newIcon) {
+  /// [ change icon in bottomSheet ]
+  changeIcon(String? newIcon) {
     selectedIcon = newIcon;
     update(['dropDownIcon', true]);
   }
 
-  // function to toggle repeat switch in BottomSheet
+  /// [ function to toggle repeat switch in BottomSheet ]
   toggleRepeat(bool newValue) {
     isRepeat = newValue;
     update(['isRepeat', true]);
   }
 
-  // convert time to minutes
+  /// [ convert time to minutes ]
   double toDouble(TimeOfDay myTime) => myTime.hour + myTime.minute / 60.0;
 
-  // function to sort allTasks based on date & time.
+  /// [ function to sort allTasks based on date & time. ]
   sortAllTasks() {
     var currDt = DateTime.now();
     allTasks.forEach((ele) {
-      if (ele.isRepeat == true) {
-        var tmp = ele.taskDate;
+      if (ele!.isRepeat == true) {
+        var tmp = ele.taskDate!;
         ele.taskDate = DateTime(
           currDt.year,
           currDt.month,
@@ -127,15 +127,15 @@ class HomeController extends GetxController {
       }
     });
     allTasks.sort((a, b) {
-      var aD = a.taskDate.toString();
-      var bD = b.taskDate.toString();
+      var aD = a!.taskDate.toString();
+      var bD = b!.taskDate.toString();
       return aD.compareTo(bD);
     });
     pastTasks.clear();
     commingTasks.clear();
-    List<Task> tmpPastTasks = [];
+    List<Task?> tmpPastTasks = [];
     allTasks.forEach((element) {
-      if (element.taskDate.isBefore(currDt)) {
+      if (element!.taskDate!.isBefore(currDt)) {
         tmpPastTasks.add(element);
       } else {
         commingTasks.add(element);
@@ -144,52 +144,52 @@ class HomeController extends GetxController {
     pastTasks = tmpPastTasks.reversed.toList();
   }
 
-  // Function to generate dailyTask
+  /// [ Function to generate dailyTask ]
   dailyTask() {
     todayTasks = [];
     var currDt = DateTime.now();
     if (allTasks.length != 0) {
       allTasks.forEach((element) {
-        if (element.taskDate.day == currDt.day &&
-            element.taskDate.month == currDt.month &&
-            element.taskDate.year == currDt.year) {
+        if (element!.taskDate!.day == currDt.day &&
+            element.taskDate!.month == currDt.month &&
+            element.taskDate!.year == currDt.year) {
           todayTasks.add(element);
         }
       });
       todayTasks.sort((a, b) {
-        var aT = toDouble(timeConvert(a.startTime));
-        var bT = toDouble(timeConvert(b.startTime));
+        var aT = toDouble(timeConvert(a!.startTime!));
+        var bT = toDouble(timeConvert(b!.startTime!));
         return aT.compareTo(bT);
       });
       todayTasks.forEach((element) {
-        print(element.startTime);
+        print(element!.startTime);
       });
       update([1, true]);
     }
   }
 
-  // function to add task via bottomSheet
+  /// [ function to add task via bottomSheet ]
   addTask() async {
     tempTask = Task();
     var box = await Hive.openBox(taskBox);
     var modDate = DateTime(
-      selectedDate.year,
-      selectedDate.month,
-      selectedDate.day,
+      selectedDate!.year,
+      selectedDate!.month,
+      selectedDate!.day,
       selectedTime.hour,
       selectedTime.minute,
     );
-    tempTask.taskImage = selectedIcon;
-    tempTask.taskTitle = titleController.text;
-    tempTask.taskDesc = descController.text;
-    tempTask.taskDate = modDate;
-    tempTask.startTime = formatDate(
+    tempTask!.taskImage = selectedIcon;
+    tempTask!.taskTitle = titleController!.text;
+    tempTask!.taskDesc = descController!.text;
+    tempTask!.taskDate = modDate;
+    tempTask!.startTime = formatDate(
         DateTime(2020, 08, 1, selectedTime.hour, selectedTime.minute),
         [hh, ':', nn, " ", am]).toString();
-    tempTask.isRepeat = isRepeat;
+    tempTask!.isRepeat = isRepeat;
     print(modDate.toString());
     allTasks.add(tempTask);
-    Map<String, dynamic> taskMap = tempTask.toJson();
+    Map<String, dynamic> taskMap = tempTask!.toJson();
     int idOfTask = await box.add(taskMap);
     taskIds = idOfTask;
     Hive.close();
@@ -199,38 +199,38 @@ class HomeController extends GetxController {
     Get.back();
   }
 
-  // preUpdates
+  /// [ preUpdates ]
   preUpdateTask(Task task) {
     selectedIcon = task.taskImage;
-    titleController.text = task.taskTitle;
-    descController.text = task.taskDesc;
-    selectedTime = timeConvert(task.startTime);
-    timeController.text = formatDate(
+    titleController!.text = task.taskTitle!;
+    descController!.text = task.taskDesc!;
+    selectedTime = timeConvert(task.startTime!);
+    timeController!.text = formatDate(
         DateTime(2020, 08, 1, selectedTime.hour, selectedTime.minute),
         [hh, ':', nn, " ", am]).toString();
     selectedDate = task.taskDate;
-    dateController.text = DateFormat.yMMMd().format(selectedDate);
+    dateController!.text = DateFormat.yMMMd().format(selectedDate!);
     isRepeat = task.isRepeat;
   }
 
-  // function to update task via bottomSheet
+  /// [ function to update task via bottomSheet ]
   updateTask(Task task) async {
     tempTask = Task();
     var modDate = DateTime(
-      selectedDate.year,
-      selectedDate.month,
-      selectedDate.day,
+      selectedDate!.year,
+      selectedDate!.month,
+      selectedDate!.day,
       selectedTime.hour,
       selectedTime.minute,
     );
-    tempTask.taskImage = selectedIcon;
-    tempTask.taskTitle = titleController.text;
-    tempTask.taskDesc = descController.text;
-    tempTask.taskDate = modDate;
-    tempTask.startTime = formatDate(
+    tempTask!.taskImage = selectedIcon;
+    tempTask!.taskTitle = titleController!.text;
+    tempTask!.taskDesc = descController!.text;
+    tempTask!.taskDate = modDate;
+    tempTask!.startTime = formatDate(
         DateTime(2020, 08, 1, selectedTime.hour, selectedTime.minute),
         [hh, ':', nn, " ", am]).toString();
-    tempTask.isRepeat = isRepeat;
+    tempTask!.isRepeat = isRepeat;
     int index = allTasks.indexOf(task);
     allTasks.setAll(index, [tempTask]);
     taskRoutine();
@@ -240,10 +240,10 @@ class HomeController extends GetxController {
     print(index);
   }
 
-  // function to read task from database
-  Future<List<Task>> getTasks() async {
+  /// [ function to read task from database ]
+  Future<List<Task?>> getTasks() async {
     var box = await Hive.openBox(taskBox);
-    List<Task> taskList = [];
+    List<Task?> taskList = [];
     for (int i = 0; i < box.length; i++) {
       var taskMap = box.getAt(i).map((k, e) => MapEntry(k.toString(), e));
       Task tmp = Task();
@@ -259,7 +259,7 @@ class HomeController extends GetxController {
     return taskList;
   }
 
-  // function to delete task by ID
+  /// [ function to delete task by ID ]
   deleteTask(Task task) async {
     int index = allTasks.indexOf(task);
     allTasks.removeAt(index);
@@ -269,27 +269,27 @@ class HomeController extends GetxController {
     print(index);
   }
 
-  // function to rewrite all tasks to DB
+  /// [ function to rewrite all tasks to DB ]
   reWriteTasks() async {
     var box = await Hive.openBox(taskBox);
     // ignore: unused_local_variable
     int id = await box.clear();
     allTasks.forEach((element) async {
-      Map<String, dynamic> newTaskMap = element.toJson();
+      Map<String, dynamic> newTaskMap = element!.toJson();
       // ignore: unused_local_variable
       int idOfTask = await box.add(newTaskMap);
     });
     Hive.close();
   }
 
-  // function to reset all controllers
+  /// [ function to reset all controllers ]
   controllerReset() {
-    titleController.text = '';
-    descController.text = '';
+    titleController!.text = '';
+    descController!.text = '';
     selectedDate = DateTime.now();
-    dateController.text = DateFormat.yMMMd().format(selectedDate);
+    dateController!.text = DateFormat.yMMMd().format(selectedDate!);
     selectedTime = TimeOfDay(hour: 00, minute: 00);
-    timeController.text = formatDate(
+    timeController!.text = formatDate(
       DateTime(2020, 08, 1, selectedTime.hour, selectedTime.minute),
       [hh, ':', nn, " ", am],
     ).toString();
@@ -297,7 +297,7 @@ class HomeController extends GetxController {
     isRepeat = false;
   }
 
-  // function to do routine on CRUD operations
+  /// [ function to do routine on CRUD operations ]
   taskRoutine() {
     sortAllTasks();
     dailyTask();
@@ -307,25 +307,25 @@ class HomeController extends GetxController {
     setTaskNotification();
   }
 
-  // setDateFunction bottomSheet
+  /// [ setDateFunction bottomSheet ]
   Future<Null> selectDate(BuildContext context) async {
-    final DateTime picked = await showDatePicker(
+    final DateTime? picked = await showDatePicker(
       context: context,
-      initialDate: selectedDate,
+      initialDate: selectedDate!,
       initialDatePickerMode: DatePickerMode.day,
       firstDate: DateTime(2015),
       lastDate: DateTime(2101),
     );
     if (picked != null) {
       selectedDate = picked;
-      dateController.text = DateFormat.yMMMd().format(selectedDate);
+      dateController!.text = DateFormat.yMMMd().format(selectedDate!);
       update();
     }
   }
 
-  // setTimeFunction bottomSheet
+  /// [ setTimeFunction bottomSheet ]
   Future<Null> selectTime(BuildContext context) async {
-    final TimeOfDay picked = await showTimePicker(
+    final TimeOfDay? picked = await showTimePicker(
       context: context,
       initialTime: selectedTime,
     );
@@ -333,16 +333,16 @@ class HomeController extends GetxController {
       selectedTime = picked;
       hour = selectedTime.hour.toString();
       minute = selectedTime.minute.toString();
-      time = hour + ' : ' + minute;
-      timeController.text = time;
-      timeController.text = formatDate(
+      time = hour! + ' : ' + minute!;
+      timeController!.text = time!;
+      timeController!.text = formatDate(
           DateTime(2020, 08, 1, selectedTime.hour, selectedTime.minute),
           [hh, ':', nn, " ", am]).toString();
       update();
     }
   }
 
-  // function to convert time from string to TimeOfDay
+  /// [ function to convert time from string to TimeOfDay ]
   TimeOfDay timeConvert(String normTime) {
     int hour;
     int minute;
@@ -362,13 +362,13 @@ class HomeController extends GetxController {
     return TimeOfDay(hour: hour, minute: minute);
   }
 
-  // function to set currentTask
+  /// [ function to set currentTask ]
   setCurrentTask() {
     var cDt = DateTime.now();
     isCurrentTaskPresent = false;
     if (todayTasks.length != 0) {
       todayTasks.forEach((element) {
-        if (element.taskDate.compareTo(cDt) <= 0) {
+        if (element!.taskDate!.compareTo(cDt) <= 0) {
           currentTask = element;
           isCurrentTaskPresent = true;
         } else {}
@@ -377,14 +377,14 @@ class HomeController extends GetxController {
     }
   }
 
-  // function to set upcomingTask
+  /// [ function to set upcomingTask ]
   setUpcomingTask() {
     isUpcommingTaskPresent = false;
     upcomingTask = Task();
     var cDt = DateTime.now();
     if (todayTasks.length != 0) {
       for (int i = 0; i < todayTasks.length; i++) {
-        if (todayTasks[i].taskDate.compareTo(cDt) > 0) {
+        if (todayTasks[i]!.taskDate!.compareTo(cDt) > 0) {
           if (isCurrentTaskPresent) {
             var index = todayTasks.indexOf(currentTask);
             upcomingTask = todayTasks.elementAt(index + 1);
@@ -399,23 +399,23 @@ class HomeController extends GetxController {
     update([3, true]);
   }
 
-  // function to get current user
+  /// [ function to get current user ]
   getUser() {
     userName = userData.read('fName');
     isMale = userData.read('isMale');
     update([7, true]);
   }
 
-  // function to set notifications for tasks.
+  /// [ function to set notifications for tasks. ]
   setTaskNotification() {
     allTasks.forEach((element) {
-      if (element.taskDate.isAfter(DateTime.now())) {
+      if (element!.taskDate!.isAfter(DateTime.now())) {
         nPlugin.showTaskNotification(element, allTasks.indexOf(element));
       }
     });
   }
 
-  // function to cancel notifications for tasks.
+  /// [ function to cancel notifications for tasks. ]
   cancellAllTaskNotification() {
     allTasks.forEach((element) {
       nPlugin.cancelTaskNotifucation(allTasks.indexOf(element));
@@ -436,11 +436,11 @@ class HomeController extends GetxController {
     descController = TextEditingController();
     dateController = TextEditingController();
     timeController = TextEditingController();
-    slideC = SlidableController();
+    // slideC = SlidableController();
     selectedDate = DateTime.now();
-    dateController.text = DateFormat.yMMMd().format(selectedDate);
+    dateController!.text = DateFormat.yMMMd().format(selectedDate!);
     selectedTime = TimeOfDay(hour: 00, minute: 00);
-    timeController.text = formatDate(
+    timeController!.text = formatDate(
         DateTime(2020, 08, 1, selectedTime.hour, selectedTime.minute),
         [hh, ':', nn, " ", am]).toString();
     selectedIcon = icons.first;
@@ -454,13 +454,13 @@ class HomeController extends GetxController {
   @override
   void onClose() {
     super.onClose();
-    titleController.dispose();
-    descController.dispose();
-    dateController.dispose();
-    timeController.dispose();
+    titleController!.dispose();
+    descController!.dispose();
+    dateController!.dispose();
+    timeController!.dispose();
   }
 
-  // function to delete confirm dialog
+  /// [ function to delete confirm dialog ]
   Future<dynamic> customDialogDel(BuildContext context, Task task) {
     return Get.dialog(Container(
       margin: EdgeInsets.symmetric(
@@ -527,7 +527,7 @@ class HomeController extends GetxController {
                         SizedBox(width: 20),
                         GestureDetector(
                           onTap: () {
-                            slideC.activeState?.close();
+                            // slideC.activeState?.close();
                             Slidable.of(context)?.close();
                             deleteTask(task);
                             Get.back();
