@@ -7,24 +7,26 @@ import 'package:taskly/app/modules/settings/controllers/settings_controller.dart
 import 'package:taskly/app/theme/app_theme.dart';
 
 class SplashScreenController extends GetxController
-    with SingleGetTickerProviderMixin {
-  // animation controller for lottie
-  AnimationController animationController;
+    // with SingleGetTickerProviderMixin {
+    with
+        GetSingleTickerProviderStateMixin {
+  /// [ animation controller for lottie ]
+  AnimationController? animationController;
 
-  // initializing userData instance
+  /// [ initializing userData instance ]
   final userData = GetStorage();
 
-  // store Login status
-  bool isLoggedIn;
+  /// [ store Login status ]
+  bool? isLoggedIn;
 
-  // function to set the user saved theme on start
+  /// [ function to set the user saved theme on start ]
   setTheme() {
     var control = Get.put(SettingsController());
     int n = control.themes.indexOf(userData.read('theme'));
     Get.changeTheme(appThemeData.values.elementAt(n));
   }
 
-  // function to check loginStatus
+  /// [ function to check loginStatus ]
   checkLogin() {
     isLoggedIn = userData.read('loginStatus');
   }
@@ -36,23 +38,23 @@ class SplashScreenController extends GetxController
     userData.writeIfNull('drinkWater', false);
     userData.writeIfNull('loginStatus', false);
 
-    // spalsh animation config
+    /// [ spalsh animation config ]
     animationController = AnimationController(
       duration: Duration(seconds: 3),
       vsync: this,
     );
-    animationController.forward();
-    animationController.addStatusListener((status) {
+    animationController!.forward();
+    animationController!.addStatusListener((status) {
       if (status == AnimationStatus.completed) {
-        animationController.reverse();
+        animationController!.reverse();
       } else if (status == AnimationStatus.dismissed) {
-        animationController.forward();
+        animationController!.forward();
       }
     });
     checkLogin();
     Timer(
       Duration(milliseconds: 5500),
-      () => Get.offNamed(isLoggedIn ? '/home' : '/welcome'),
+      () => Get.offNamed(isLoggedIn! ? '/home' : '/welcome'),
     );
   }
 
@@ -64,7 +66,7 @@ class SplashScreenController extends GetxController
 
   @override
   void onClose() async {
-    animationController.dispose();
+    animationController!.dispose();
     super.onClose();
   }
 }
