@@ -3,25 +3,26 @@ import 'package:flutter/material.dart';
 final white1 = Colors.white;
 final white2 = Color(0xffeeeeee);
 
-// Define yellow colors here
+///  [ Define yellow colors here ]
 final primaryYellow = Color(0xffffc045);
 final secondaryYellow = Color(0xfffdb827);
 final secondaryGrey = Color(0xff3a4750);
 final primaryGrey = Color(0xff303841);
 
-// Define red colors here
+/// [  Define red colors here ]
 final primaryRed = Color(0xfff05454);
 final secondaryDBlue = Color(0xff3a4750);
 final primaryDBlue = Color(0xff222831);
 
-// Define teal colors here
+/// [ Define teal colors here ]
 final primaryTeal = Color(0xff00adb5);
 final secondaryDGrey = Color(0xff393e46);
 final primaryDGrey = Color(0xff222831);
 
-// Define green colors here
+/// [ Define green colors here ]
 final primaryGreen = Color(0xff7bc74d);
 final secondaryBGrey = Color(0xff393e46);
 final primaryBGrey = Color(0xff222831);
 
+/// [ Define hint color here ]
 final authHintText = Color(0xff8E8E8E);
